+++ src/UserInfoBlock/index.js
@@ -4,31 +4,30 @@ import {Link} from "react-router-dom";
 
 const userRepositoryCount = "User repository count ";
 
-class UserComponent extends Component {
+class UserInfo extends Component {
     render() {
+        const {avatar, username, signOut} = this.props;
         return (
             <div className="card-300 mb-4">
                 <Card>
                     <div className="card card-inverse">
                         <div className="card-img-overlay">
                             <div className="flex justify-content-between">
-                                <Link to='/user_settings'>
+                                <Link to='/settings'>
                                     <button type="button" className="btn btn-light btn-sm">Settings</button>
                                 </Link>
                                 <Link to='/'>
-                                    <button type="button" onClick={this.props.signOut}
-                                            className="btn btn-light btn-sm">Sign out
+                                    <button type="button" onClick={signOut} className="btn btn-light btn-sm">Sign out
                                     </button>
                                 </Link>
                             </div>
                         </div>
                     </div>
-                    <CardImg top width="100%" src="https://picsum.photos/300/180" alt="Card image cap"/>
+                    <CardImg top width="100%" src="https://picsum.photos/300/180?image=1045" alt=""/>
                     <CardBody>
-                        <Link to='/user'><img
-                            src={this.props.avatar ? this.props.avatar : 'https://picsum.photos/150/150'}
-                            alt="Avatar" className="br-50"/></Link>
-                        <CardTitle>{this.props.username}</CardTitle>
+                        <Link to='/'><img src={avatar ? avatar : 'https://picsum.photos/150/150'}
+                                          alt="Avatar" className="br-50"/></Link>
+                        <CardTitle>{username}</CardTitle>
                         <CardText>{userRepositoryCount}</CardText>
                     </CardBody>
                 </Card>
@@ -38,4 +37,4 @@ class UserComponent extends Component {
     }
 }
 
-export default UserComponent;
+export default UserInfo;
