+++ src/_components/UserSettings.js
@@ -1,7 +1,7 @@
 import React, {Component} from 'react';
-import Search from '../HomePage/search'
+import Search from '../UserPage/search'
 import Content from './content';
-import UserComponent from './user-component';
+import UserComponent from '../UserInfoBlock/index';
 import {Button} from 'reactstrap';
 import {Link} from 'react-router-dom';
 import UserSettings from "./_components/user-settings";
